# Removed loops over batch sizes