Changes:
- Loop over batch dimension was replaced by putting the batch dimension
  into the y dimension of the grid. This limits the maximum possible
  batch size to 2**16.
- Changed block and grid dimensions to prevent unnecessarily large
  blocks that do nothing.
- Removed small limit in grid-sizes in x-direction (MAX_PIXELS_2d,
  MAX_PIXELS_3d). The hardware limit is 2**32 -1 = 4294967295, which is
  way higher. If there was a reason for that limit then I didn't see it.
- Added comments to the kernel code. This is a side effect of a
  debugging session.

Effects:
For large batches of small images the utilization is now way higher.
For the denoise.py test one epoch is much faster.
E.g. On a TITAN X:
torch: 237.0s
old: 160.0s
new: 50.3s

Timing comparison with the test/denoise.py highlight the performance gain of the changes:

| Old kernels Training Stats  |  New kernels Training Stats | Old Evaluation Loss PSNR | New Evaluation Loss PSNR |
| ----------------------------|-----------------------------|-----------------------|-----------------------------|
| Epoch 0:	0.00740	158.7s  |  Epoch 0:	0.00739	50.1s| 0.00737	21.324              |  0.00736	21.332|
| Epoch 1:	0.00593	159.9s  |  Epoch 1:	0.00686	52.7s| 0.00589	22.299              |  0.00683	21.653|
| Epoch 2:	0.00462	159.9s  |  Epoch 2:	0.00613	54.1s| 0.00460	23.376              |  0.00610	22.145|
| Epoch 3:	0.00412	160.0s  |  Epoch 3:	0.00524	54.5s| 0.00411	23.861              |  0.00519	22.849|
| Epoch 4:	0.00396	159.9s  |  Epoch 4:	0.00456	54.3s| 0.00396	24.022              |  0.00454	23.429|
| Epoch 5:	0.00381	160.0s  |  Epoch 5:	0.00417	53.9s| 0.00382	24.179              |  0.00417	23.797|
| Epoch 6:	0.00370	159.9s  |  Epoch 6:	0.00397	53.7s| 0.00368	24.345              |  0.00397	24.011|
| Epoch 7:	0.00363	160.0s  |  Epoch 7:	0.00389	53.5s| 0.00363	24.402              |  0.00402	23.954|
| Epoch 8:	0.00363	160.1s  |  Epoch 8:	0.00380	53.5s| 0.00359	24.449              |  0.00381	24.194|
| Epoch 9:	0.00351	160.0s  |  Epoch 9:	0.00371	53.5s| 0.00353	24.519              |  0.00393	24.058|
| Epoch 10:	0.00350	160.0s  |  Epoch 10:	0.00367	53.5s| 0.00348	24.589              |  0.00364	24.384|
| Epoch 11:	0.00347	160.0s  |  Epoch 11:	0.00360	53.4s| 0.00349	24.578              |  0.00360	24.433|
| Epoch 12:	0.00345	160.0s  |  Epoch 12:	0.00354	53.4s| 0.00342	24.656              |  0.00356	24.487|
| Epoch 13:	0.00341	160.1s  |  Epoch 13:	0.00349	53.4s| 0.00339	24.692              |  0.00350	24.565|
| Epoch 14:	0.00341	160.1s  |  Epoch 14:	0.00344	53.5s| 0.00341	24.668              |  0.00343	24.649|
| Epoch 15:	0.00336	160.0s  |  Epoch 15:	0.00340	53.4s| 0.00335	24.751              |  0.00340	24.680|
| Epoch 16:	0.00335	160.0s  |  Epoch 16:	0.00337	53.5s| 0.00338	24.706              |  0.00335	24.746|
| Epoch 17:	0.00333	160.0s  |  Epoch 17:	0.00339	53.5s| 0.00334	24.765              |  0.00342	24.659|
| Epoch 18:	0.00333	160.0s  |  Epoch 18:	0.00335	53.6s| 0.00330	24.816              |  0.00344	24.633|
| Epoch 19:	0.00331	160.0s  |  Epoch 19:	0.00331	53.5s| 0.00331	24.800              |  0.00333	24.771|
| Epoch 20:	0.00329	160.0s  |  Epoch 20:	0.00329	53.4s| 0.00338	24.715              |  0.00328	24.846|
| Epoch 21:	0.00327	160.0s  |  Epoch 21:	0.00331	53.3s| 0.00332	24.786              |  0.00328	24.844|
| Epoch 22:	0.00327	160.0s  |  Epoch 22:	0.00328	53.3s| 0.00328	24.843              |  0.00328	24.846|
| Epoch 23:	0.00321	160.0s  |  Epoch 23:	0.00322	53.3s| 0.00321	24.942              |  0.00321	24.930|
| Epoch 24:	0.00323	160.2s  |  Epoch 24:	0.00324	53.3s| 0.00320	24.949              |  0.00321	24.935|
| Epoch 25:	0.00320	160.0s  |  Epoch 25:	0.00321	53.5s| 0.00320	24.953              |  0.00321	24.942|
| Epoch 26:	0.00321	160.0s  |  Epoch 26:	0.00322	53.3s| 0.00320	24.950              |  0.00321	24.941|
| Epoch 27:	0.00321	160.0s  |  Epoch 27:	0.00320	53.4s| 0.00320	24.947              |  0.00320	24.945|
| Epoch 28:	0.00319	160.0s  |  Epoch 28:	0.00319	53.2s| 0.00319	24.962              |  0.00321	24.938|
| Epoch 29:	0.00322	160.0s  |  Epoch 29:	0.00323	53.2s| 0.00319	24.957              |  0.00321	24.936|
| Epoch 30:	0.00320	160.0s  |  Epoch 30:	0.00321	53.2s| 0.00320	24.952              |  0.00319	24.957|
| Epoch 31:	0.00319	160.0s  |  Epoch 31:	0.00322	53.2s| 0.00318	24.969              |  0.00319	24.956|
| Epoch 32:	0.00319	160.0s  |  Epoch 32:	0.00320	53.2s| 0.00319	24.963              |  0.00319	24.958|
| Epoch 33:	0.00318	160.0s  |  Epoch 33:	0.00320	53.3s| 0.00318	24.974              |  0.00319	24.968|
| Epoch 34:	0.00319	160.0s  |  Epoch 34:	0.00320	53.3s| 0.00318	24.970              |  0.00318	24.972|
| Epoch 35:	0.00318	160.1s  |  Epoch 35:	0.00320	53.3s| 0.00318	24.980              |  0.00319	24.966|
| Epoch 36:	0.00320	160.2s  |  Epoch 36:	0.00320	53.3s| 0.00318	24.982              |  0.00318	24.969|
| Epoch 37:	0.00318	160.0s  |  Epoch 37:	0.00318	53.3s| 0.00318	24.982              |  0.00318	24.979|
| Epoch 38:	0.00319	160.0s  |  Epoch 38:	0.00318	53.3s| 0.00317	24.987              |  0.00318	24.976|
| Epoch 39:	0.00320	160.0s  |  Epoch 39:	0.00320	53.2s| 0.00319	24.968              |  0.00318	24.978|

Note that usage of shared memory might still help to significantly improve performance further. We might commit that in a future pull request.